Upgrade to WPT V 2.13

+++ xmlResult.php
@@ -133,16 +133,18 @@ else
               $key = preg_replace('/[^a-zA-Z0-9\.\-_]/', '_', $key);
               echo "<$key>" . xml_entities($val) . "</$key>\n";
             }
-            if( $pagespeed )
-            {
-                $score = GetPageSpeedScore("$testPath/{$fvMedian}_pagespeed.txt");
-                if( strlen($score) )
-                    echo "<PageSpeedScore>$score</PageSpeedScore>\n";
+            if (gz_is_file("$testPath/{$fvMedian}_pagespeed.txt")) {
+              if( $pagespeed )
+              {
+                  $score = GetPageSpeedScore("$testPath/{$fvMedian}_pagespeed.txt");
+                  if( strlen($score) )
+                      echo "<PageSpeedScore>$score</PageSpeedScore>\n";
+              }
+              if( FRIENDLY_URLS )
+                  echo "<PageSpeedData>http://$host$uri/result/$id/{$fvMedian}_pagespeed.txt</PageSpeedData>\n";
+              else
+                  echo "<PageSpeedData>http://$host$uri//getgzip.php?test=$id&amp;file={$fvMedian}_pagespeed.txt</PageSpeedData>\n";
             }
-            if( FRIENDLY_URLS )
-                echo "<PageSpeedData>http://$host$uri/result/$id/{$fvMedian}_pagespeed.txt</PageSpeedData>\n";
-            else
-                echo "<PageSpeedData>http://$host$uri//getgzip.php?test=$id&amp;file={$fvMedian}_pagespeed.txt</PageSpeedData>\n";
             xmlDomains($id, $testPath, $fvMedian, 0);
             xmlBreakdown($id, $testPath, $fvMedian, 0);
             xmlRequests($id, $testPath, $fvMedian, 0);
@@ -161,16 +163,18 @@ else
                       $key = preg_replace('/[^a-zA-Z0-9\.\-_]/', '_', $key);
                       echo "<$key>" . xml_entities($val) . "</$key>\n";
                     }
-                    if( $pagespeed )
-                    {
-                        $score = GetPageSpeedScore("$testPath/{$rvMedian}_Cached_pagespeed.txt");
-                        if( strlen($score) )
-                            echo "<PageSpeedScore>$score</PageSpeedScore>\n";
+                    if (gz_is_file("$testPath/{$fvMedian}_Cached_pagespeed.txt")) {
+                      if( $pagespeed )
+                      {
+                          $score = GetPageSpeedScore("$testPath/{$rvMedian}_Cached_pagespeed.txt");
+                          if( strlen($score) )
+                              echo "<PageSpeedScore>$score</PageSpeedScore>\n";
+                      }
+                      if( FRIENDLY_URLS )
+                          echo "<PageSpeedData>http://$host$uri/result/$id/{$rvMedian}_Cached_pagespeed.txt</PageSpeedData>\n";
+                      else
+                          echo "<PageSpeedData>http://$host$uri//getgzip.php?test=$id&amp;file={$rvMedian}_Cached_pagespeed.txt</PageSpeedData>\n";
                     }
-                    if( FRIENDLY_URLS )
-                        echo "<PageSpeedData>http://$host$uri/result/$id/{$rvMedian}_Cached_pagespeed.txt</PageSpeedData>\n";
-                    else
-                        echo "<PageSpeedData>http://$host$uri//getgzip.php?test=$id&amp;file={$rvMedian}_Cached_pagespeed.txt</PageSpeedData>\n";
                     xmlDomains($id, $testPath, $fvMedian, 1);
                     xmlBreakdown($id, $testPath, $fvMedian, 1);
                     xmlRequests($id, $testPath, $fvMedian, 1);
@@ -231,52 +235,65 @@ else
                     echo "<thumbnails>\n";
                     echo "<waterfall>http://$host$uri/result/$id/{$i}_waterfall_thumb.png</waterfall>\n";
                     echo "<checklist>http://$host$uri/result/$id/{$i}_optimization_thumb.png</checklist>\n";
-                    echo "<screenShot>http://$host$uri/result/$id/{$i}_screen_thumb.jpg</screenShot>\n";
+                    if( is_file("$testPath/{$i}_screen.jpg") )
+                      echo "<screenShot>http://$host$uri/result/$id/{$i}_screen_thumb.jpg</screenShot>\n";
                     echo "</thumbnails>\n";
 
                     echo "<images>\n";
                     echo "<waterfall>http://$host$uri$path/{$i}_waterfall.png</waterfall>\n";
                     echo "<connectionView>http://$host$uri$path/{$i}_connection.png</connectionView>\n";
                     echo "<checklist>http://$host$uri$path/{$i}_optimization.png</checklist>\n";
-                    echo "<screenShot>http://$host$uri$path/{$i}_screen.jpg</screenShot>\n";
+                    if( is_file("$testPath/{$i}_screen.jpg") )
+                      echo "<screenShot>http://$host$uri$path/{$i}_screen.jpg</screenShot>\n";
                     if( is_file("$testPath/{$i}_screen.png") )
                         echo "<screenShotPng>http://$host$uri$path/{$i}_screen.png</screenShotPng>\n";
                     echo "</images>\n";
 
                     // raw results
                     echo "<rawData>";
-                    echo "<headers>http://$host$uri$path/{$i}_report.txt</headers>\n";
-                    if (array_key_exists('bodies', $test['testinfo']) && $test['testinfo']['bodies']) {
+                    if (gz_is_file("$testPath/{$i}_report.txt"))
+                      echo "<headers>http://$host$uri$path/{$i}_report.txt</headers>\n";
+                    if (is_file("$testPath/{$i}_bodies.zip"))
                         echo "<bodies>http://$host$uri$path/{$i}_bodies.zip</bodies>\n";
-                    }
-                    echo "<pageData>http://$host$uri$path/{$i}_IEWPG.txt</pageData>\n";
-                    echo "<requestsData>http://$host$uri$path/{$i}_IEWTR.txt</requestsData>\n";
-                    echo "<utilization>http://$host$uri$path/{$i}_progress.csv</utilization>\n";
-                    echo "<PageSpeedData>http://$host$uri/result/$id/{$i}_pagespeed.txt</PageSpeedData>\n";
+                    if (gz_is_file("$testPath/{$i}_IEWPG.txt"))
+                      echo "<pageData>http://$host$uri$path/{$i}_IEWPG.txt</pageData>\n";
+                    if (gz_is_file("$testPath/{$i}_IEWTR.txt"))
+                      echo "<requestsData>http://$host$uri$path/{$i}_IEWTR.txt</requestsData>\n";
+                    if (gz_is_file("$testPath/{$i}_progress.csv"))
+                      echo "<utilization>http://$host$uri$path/{$i}_progress.csv</utilization>\n";
+                    if (gz_is_file("$testPath/{$i}_pagespeed.txt"))
+                      echo "<PageSpeedData>http://$host$uri/result/$id/{$i}_pagespeed.txt</PageSpeedData>\n";
                     echo "</rawData>\n";
                     
                     // video frames
-                    if( (array_key_exists('video', $test['test']) && $test['test']['video']) ||
-                        (array_key_exists('video', $test['testinfo']) && $test['testinfo']['video']) )
-                    {
-                        loadVideo("$testPath/video_{$i}", $frames);
-                        if( $frames && count($frames) )
-                        {
-                            $progress = GetVisualProgress($testPath, $i, 0);
-                            echo "<videoFrames>\n";
-                            foreach( $frames as $time => $frameFile )
-                            {
-                                echo "<frame>\n";
-                                echo "<time>" . number_format((double)$time / 10.0, 1) . "</time>\n";
-                                echo "<image>http://$host$uri$path/video_{$i}/$frameFile</image>\n";
-                                $ms = $time * 100;
-                                if (isset($progress) && is_array($progress) && 
-                                    array_key_exists('frames', $progress) && array_key_exists($ms, $progress['frames'])) {
-                                    echo "<VisuallyComplete>{$progress['frames'][$ms]['progress']}</VisuallyComplete>\n";
-                                }
-                                echo "</frame>\n";
-                            }
-                            echo "</videoFrames>\n";
+                    $startOffset = array_key_exists('testStartOffset', $pageData[$i][0]) ? intval(round($pageData[$i][0]['testStartOffset'])) : 0;
+                    $progress = GetVisualProgress($testPath, $i, 0, null, null, $startOffset);
+                    if (array_key_exists('frames', $progress) && is_array($progress['frames']) && count($progress['frames'])) {
+                      echo "<videoFrames>\n";
+                      foreach($progress['frames'] as $ms => $frame) {
+                          echo "<frame>\n";
+                          echo "<time>$ms</time>\n";
+                          echo "<image>http://$host$uri$path/video_{$i}/{$frame['file']}</image>\n";
+                          echo "<VisuallyComplete>{$frame['progress']}</VisuallyComplete>\n";
+                          echo "</frame>\n";
+                      }
+                      echo "</videoFrames>\n";
+                    }
+                    if (isset($progress) &&
+                        is_array($progress) &&
+                        array_key_exists('DevTools', $progress) &&
+                        is_array($progress['DevTools'])) {
+                        if (array_key_exists('processing', $progress['DevTools'])) {
+                          echo "<processing>\n";
+                          foreach ($progress['DevTools']['processing'] as $key => $value)
+                            echo "<$key>$value</$key>\n";
+                          echo "</processing>\n";
+                        }
+                        if (array_key_exists('VisualProgress', $progress['DevTools'])) {
+                          echo "<VisualProgress>\n";
+                          foreach ($progress['DevTools']['VisualProgress'] as $key => $value)
+                            echo "<progress>\n<time>$key</time>\n<VisuallyComplete>$value</VisuallyComplete>\n</progress>\n";
+                          echo "</VisualProgress>\n";
                         }
                     }
                     
@@ -319,53 +336,59 @@ else
                     echo "<thumbnails>\n";
                     echo "<waterfall>http://$host$uri/result/$id/{$i}_Cached_waterfall_thumb.png</waterfall>\n";
                     echo "<checklist>http://$host$uri/result/$id/{$i}_Cached_optimization_thumb.png</checklist>\n";
-                    echo "<screenShot>http://$host$uri/result/$id/{$i}_Cached_screen_thumb.jpg</screenShot>\n";
+                    if( is_file("$testPath/{$i}_Cached_screen.jpg") )
+                      echo "<screenShot>http://$host$uri/result/$id/{$i}_Cached_screen_thumb.jpg</screenShot>\n";
                     echo "</thumbnails>\n";
 
                     echo "<images>\n";
                     echo "<waterfall>http://$host$uri$path/{$i}_Cached_waterfall.png</waterfall>\n";
                     echo "<connectionView>http://$host$uri$path/{$i}_Cached_connection.png</connectionView>\n";
                     echo "<checklist>http://$host$uri$path/{$i}_Cached_optimization.png</checklist>\n";
-                    echo "<screenShot>http://$host$uri$path/{$i}_Cached_screen.jpg</screenShot>\n";
+                    if( is_file("$testPath/{$i}_Cached_screen.jpg") )
+                      echo "<screenShot>http://$host$uri$path/{$i}_Cached_screen.jpg</screenShot>\n";
                     if( is_file("$testPath/{$i}_Cached_screen.png") )
                         echo "<screenShotPng>http://$host$uri$path/{$i}_Cached_screen.png</screenShotPng>\n";
                     echo "</images>\n";
 
                     // raw results
                     echo "<rawData>\n";
-                    echo "<headers>http://$host$uri$path/{$i}_Cached_report.txt</headers>\n";
-                    if (array_key_exists('bodies', $test['testinfo']) && $test['testinfo']['bodies']) {
+                    if (gz_is_file("$testPath/{$i}_Cached_report.txt"))
+                      echo "<headers>http://$host$uri$path/{$i}_Cached_report.txt</headers>\n";
+                    if (is_file("$testPath/{$i}_Cached_bodies.zip"))
                         echo "<bodies>http://$host$uri$path/{$i}_Cached_bodies.zip</bodies>\n";
-                    }
-                    echo "<pageData>http://$host$uri$path/{$i}_Cached_IEWPG.txt</pageData>\n";
-                    echo "<requestsData>http://$host$uri$path/{$i}_Cached_IEWTR.txt</requestsData>\n";
-                    echo "<utilization>http://$host$uri$path/{$i}_Cached_progress.csv</utilization>\n";
-                    echo "<PageSpeedData>http://$host$uri/result/$id/{$i}_Cached_pagespeed.txt</PageSpeedData>\n";
+                    if (gz_is_file("$testPath/{$i}_Cached_IEWPG.txt"))
+                      echo "<pageData>http://$host$uri$path/{$i}_Cached_IEWPG.txt</pageData>\n";
+                    if (gz_is_file("$testPath/{$i}_Cached_IEWTR.txt"))
+                      echo "<requestsData>http://$host$uri$path/{$i}_Cached_IEWTR.txt</requestsData>\n";
+                    if (gz_is_file("$testPath/{$i}_Cached_progress.csv"))
+                      echo "<utilization>http://$host$uri$path/{$i}_Cached_progress.csv</utilization>\n";
+                    if (gz_is_file("$testPath/{$i}_Cached_pagespeed.txt"))
+                      echo "<PageSpeedData>http://$host$uri/result/$id/{$i}_Cached_pagespeed.txt</PageSpeedData>\n";
                     echo "</rawData>\n";
                     
                     // video frames
-                    if( (array_key_exists('video', $test['test']) && $test['test']['video']) ||
-                        (array_key_exists('video', $test['testinfo']) && $test['testinfo']['video']) )
-                    {
-                        loadVideo("$testPath/video_{$i}_cached", $frames);
-                        if( $frames && count($frames) )
-                        {
-                            $progress = GetVisualProgress($testPath, $i, 1);
-                            echo "<videoFrames>\n";
-                            foreach( $frames as $time => $frameFile )
-                            {
-                                echo "<frame>\n";
-                                echo "<time>" . number_format((double)$time / 10.0, 1) . "</time>\n";
-                                echo "<image>http://$host$uri$path/video_{$i}_cached/$frameFile</image>\n";
-                                $ms = $time * 100;
-                                if (isset($progress) && is_array($progress) && 
-                                    array_key_exists('frames', $progress) && array_key_exists($ms, $progress['frames'])) {
-                                    echo "<VisuallyComplete>{$progress['frames'][$ms]['progress']}</VisuallyComplete>\n";
-                                }
-                                echo "</frame>\n";
-                            }
-                            echo "</videoFrames>\n";
-                        }
+                    $startOffset = array_key_exists('testStartOffset', $pageData[$i][1]) ? intval(round($pageData[$i][1]['testStartOffset'])) : 0;
+                    $progress = GetVisualProgress($testPath, $i, 1, null, null, $startOffset);
+                    if (array_key_exists('frames', $progress) && is_array($progress['frames']) && count($progress['frames'])) {
+                      echo "<videoFrames>\n";
+                      foreach($progress['frames'] as $ms => $frame) {
+                          echo "<frame>\n";
+                          echo "<time>$ms</time>\n";
+                          echo "<image>http://$host$uri$path/video_{$i}_cached/{$frame['file']}</image>\n";
+                          echo "<VisuallyComplete>{$frame['progress']}</VisuallyComplete>\n";
+                          echo "</frame>\n";
+                      }
+                      echo "</videoFrames>\n";
+                    }
+                    if (isset($progress) &&
+                        is_array($progress) &&
+                        array_key_exists('DevTools', $progress) &&
+                        is_array($progress['DevTools']) &&
+                        array_key_exists('processing', $progress['DevTools'])) {
+                        echo "<processing>\n";
+                        foreach ($progress['DevTools']['processing'] as $key => $value)
+                          echo "<$key>$value</$key>\n";
+                        echo "</processing>\n";
                     }
                     
                     xmlDomains($id, $testPath, $i, 1);
@@ -394,6 +417,7 @@ else
         echo "<response>\n";
         if( strlen($_REQUEST['r']) )
             echo "<requestId>{$_REQUEST['r']}</requestId>\n";
+
         // see if it was a valid test
         if( $test['test']['runs'] )
         {
@@ -556,26 +580,6 @@ function ConsoleLog($id, $testPath, $run, $cached) {
         }
         echo "</consoleLog>\n";
     }
-
-/*    -    $cachedText = '';
--    if ($cached)
--        $cachedText = '_Cached';
--    $consoleLogFile = "$testPath/$run{$cachedText}_console_log.json";
--    if (gz_is_file($consoleLogFile)) {
--        $consoleLog = json_decode(gz_file_get_contents($consoleLogFile), true);
--        if (isset($consoleLog) && is_array($consoleLog) && count($consoleLog)) {
--            echo "<consoleLog>\n";
--            foreach( $consoleLog as &$entry ) {
--                echo "<entry>\n";
--                echo "<source>" . xml_entities($entry['source']) . "</source>\n";
--                echo "<level>" . xml_entities($entry['level']) . "</level>\n";
--                echo "<message>" . xml_entities($entry['text']) . "</message>\n";
--                echo "<url>" . xml_entities($entry['url']) . "</url>\n";
--                echo "<line>" . xml_entities($entry['line']) . "</line>\n";
--                echo "</entry>\n";
--            }
--            echo "</consoleLog>\n";
-*/
 }
 
 /**
